Create Django Natural Disaster Management
This Natural Disaster Management System Project in Django created based on python, Django, and SQLITE3 Database. A disaster management system (DMS) is a database that collects and reports data about natural disasters. In the event of a natural disaster, residents in the affected region can report the incidence. This occurrence can be recorded in a database and reported to government officials so that proper accident prevention measures can be taken.

A Natural Disaster Management System in Django is an easy project for beginners to learn how to build a web-based python Django project. We will provide you with the complete source code and database for the python project so that you can easily install it on your machine and learn how to program in Python Django.
To start creating a Natural Disaster Management System Project in Python Django, makes sure that you have PyCharm Professional IDE Installed in your computer.