Add files via upload
This Wedding Management System in Django created based on python, Django, and SQLITE3 Database. The Party model allows you to gather your guests into groups for purposes such as sending a single invitation to a pair.
You may also use the is invited field to include parties you’re not sure you’ll invite, which is useful for sending tiered invitations. There’s also a field for keeping note of whether or not the party has been invited to the rehearsal dinner.
An Wedding Management System in Django is an easy project for beginners to learn how to build a web-based python Django project. We will provide you with the complete source code and database for the python project so that you can easily install it on your machine and learn how to program in Python Django.
To start creating a Wedding Management System Project in Python Django, makes sure that you have PyCharm Professional IDE Installed in your computer.
This Wedding Management System in Django Framework, Also includes a Download Source Code for free, just find the downloadable source code below and click download now.